requirements.txt
pyglet
pygame
opencv-python>=4.7
box2d-py

tqdm
gymnasium